random fixes

+++ include/server_clib/rnd.h
@@ -10,7 +10,7 @@ extern "C" {
 
 uint64_t create_pseudo_random(const uint64_t seed, const uint64_t offset);
 
-uint32_t create_pseudo_random_from_time(const uint32_t offset);
+uint64_t create_pseudo_random_from_time(const uint64_t offset);
 
 #ifdef __cplusplus
 }
