+++ src/rnd.c
@@ -14,7 +14,7 @@ uint64_t create_pseudo_random(const uint64_t seed, const uint64_t offset)
     return r;
 }
 
-uint32_t create_pseudo_random_from_time(const uint32_t offset)
+uint64_t create_pseudo_random_from_time(const uint64_t offset)
 {
     time_t now = time(NULL);
